pyproject.toml
[tool.poetry]
name = "mention2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pymongo = "^3.11.0"
pyrogram = "^0.18.0"
aiogram = "^2.9.2"
aiohttp = "^3.6.2"
bson = "^0.5.10"
more-itertools = "^8.4.0"
dnspython = "^1.16.0"
TgCrypto = "^1.2.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
